+++ src/app/page.tsx
@@ -2,7 +2,7 @@
 import { Suspense } from 'react'
 import { hasCookie } from 'cookies-next'
 import { BottomNavbarContainer } from './_components/Common/BottomNavbar'
-import Navbar from './_components/Common/Navbar/NavigationBar'
+import FeedNavbar from './_components/Common/FeedNavbar/FeedNavigationBar'
 import Loading from './loading'
 import { PostCardList } from './post/_components/CardList/PostCardList'
 
@@ -12,7 +12,7 @@ export default function Home({ searchParams }: { searchParams: { page: string } 
 
   return (
     <div>
-      <Navbar isLogin={isLogin} />
+      <FeedNavbar isLogin={isLogin} />
       <main className="flex items-center justify-center py-20">
         <Suspense fallback={<Loading />}>
           <PostCardList pageId={pageId} />
