+++ src/app/post/[postId]/_components/PostDetail/PostDetail.tsx
@@ -9,7 +9,7 @@ import ShopInfo from '@/app/post/_components/CardItem/ShopInfo/ShopInfo'
 import { useGetPostDetail } from '@/app/post/hooks/useGetPostDetail'
 import { Button } from '@/components/ui/button'
 import { toast } from '@/components/ui/use-toast'
-import { BackNavbar } from '../../../../_components/Common/Navbar/BackNavigationBar'
+import { BackNavbar } from '../../../../_components/Common/BackNavbar/BackNavigationBar'
 import { PostHeader } from '../../../_components/CardItem/Header'
 import ImageOrModelViewer from '../MediaToggleViewer/MediaToggleViewer'
 import Rating from '../Rating/Rating'
