+++ src/app/post/[postId]/edit/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { BackNavbar } from '@/app/_components/Common/Navbar/BackNavigationBar'
+import { BackNavbar } from '@/app/_components/Common/BackNavbar/BackNavigationBar'
 import Loading from '@/app/loading'
 import { PostDetailType } from '@/types/Post/types'
 import { useGetPostDetail } from '../../hooks/useGetPostDetail'
