+++ src/app/shop/search/page.tsx
@@ -1,7 +1,7 @@
 // SearchShopPage.tsx
 import { Suspense } from 'react'
+import { BackNavbar } from '@/app/_components/Common/BackNavbar/BackNavigationBar'
 import { BottomNavbarContainer } from '@/app/_components/Common/BottomNavbar'
-import { BackNavbar } from '@/app/_components/Common/Navbar/BackNavigationBar'
 import Loading from '@/app/loading'
 import ShopSearchInput from './_components/Search/ShopSearchInput'
 import ShopList from './_components/ShopList'
